Add Microsoft OpenTelemetry distro and refactor AzMon distro
